Update microsite/src/pages/on-demand/_onDemandCard.tsx: The code change in this git diff is about adding a new prop to the OnDemandCard component in React.
1. A new prop named 'children' is added to the prop type definition of the OnDemandCard component. This prop is optional and its type is React.ReactNode. ReactNode includes all types that a render method can return, so it can be a string, a number, a React element, an array, or even a function.

```typescript
children?: React.ReactNode;
```

This change is made to explicitly list the 'children' prop when defining props. This is useful when you want to render some content inside the component and you don't know what that content will be ahead of time. It could be text, another React component, or nothing at all. By adding 'children' to the prop types, you're telling TypeScript that this component can accept children and they can be of any type that a render method can return.